Refactor: Moving generic board code from GoEngine to Board
